#3 V1.0.0 Pedestrian Moving
This release temporarily resolve the issue #3.  Change has been made as follow;
Since the OS handles key repeats, holding down a key may cause multiple calls to keyPressed(), whereas keyReleased() will only be call once when a specific key is released. Control structure for vertical moving will execute when keyReleased() is called instead of keyPressed() to avoid repeating key-events.
Outstanding issue: KeyReleased() will only be called when a key is released which result in delayed response when a move key is triggered.
This issue is progressively refined in each release in SppedFixup.