START Multiple vehicles in multiple lanes
A major update on entire game design. Using 2 dimensional array to replace former single X and Y position. This allows vehicles to spawn at different X and Y positions.
A 2d array for vehicle velocity is also assigned to give each vehicle a dynamic velocity which can be manipulate by its acceleration in the next update.
The vehicle reset feature is also redesigned and a exclusive function is created for it. The reset function is now capable to regenerate vehicles in any specific lane when last vehicle in that lane passes right boundary.
Draw vehicle function is now able to fetch X and Y positions from those arrays and draw the vehicles in different positions.